python_lh3__api_server/pyproject.toml
[tool.poetry]
name = "python_lh3__api_server"
version = "0.1.0"
description = ""
authors = ["Guinsly Mondesir <guinslym@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
msgpack-python = "^0.5.6"
pytest-sugar = "^0.9.3"
flask = "^1.1.2"
gunicorn = "^20.0.4"
cherrypy = "^18.6.0"
lh3api = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
